nyota-enterprise/growth/requirements.txt
fastapi
uvicorn
nats-py
pydantic
psycopg2-binary
playwright
beautifulsoup4
httpx
openai
qdrant-client
fastembed
